.history/requirements_20250908183830.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
pydantic==1.10.9
python-dateutil==2.8.2
Jinja2==3.1.2